requirements.txt
openai-whisper>=20231117
torch>=2.7.1
torchvision>=0.22.1
torchaudio>=2.7.1
PySide6>=6.9.1
ffmpeg-python>=0.2.0
numpy>=2.0.0